[FIX] base: fix search calls not always returning the correct record count or number of records
The ORM search function in some cases produces the wrong number of records when a limit is applied.
This is because the limit is applied within the SQL query, but the SQL output was 'uniqified' afterwards.
This commit makes the sql query be constructed with DISTINCT, so that the limit may be applied afterwards.